creates a basepage and modify login tests
