+++ pages/login.page.js
@@ -1,15 +1,13 @@
 const PAGE_TIMEOUT = 10000;
+
 class LoginPage {
     
     get usernameInput () { return $('#username'); }
     get passwordInput () { return $('#password'); }
     get signInButton () { return $('#submit'); }
     get rememberMe () { return $('#remember-me'); }
+    get alert () {return $('[class= "badge badge-pill badge-danger"]'); }
 
-
-    async open(link) {
-        await browser.url(`${link}`);
-    }
 
     async logIn(username, password) {
         await this.usernameInput.waitForClickable({ timeout: PAGE_TIMEOUT });
@@ -18,11 +16,5 @@ class LoginPage {
         await this.passwordInput.setValue(password);
     }
 
-    async clickElement (element) {
-        await element.waitForClickable({ timeout: PAGE_TIMEOUT });
-        await element.click();
-    }
 }
-
-
 export default new LoginPage();
