+++ tests/login.test.js
@@ -1,57 +1,56 @@
+import BasePage from '../pages/base.page';
 import LoginPage from '../pages/login.page';
 
 describe('DigitalBank', () => {
 
     //example test
     it('Should display login page when browsing to the route', async () => {
-      await LoginPage.open('login');
-     // expect(browser.title().value()).toMatch('login-logo');
+      await BasePage.open('login');
+      expect(browser).toHaveUrlContaining('login');
     });
 
     it('Should log into demo account when valid data is entered', async () => {
      
-      await LoginPage.open('login');
+      await BasePage.open('login');
       await LoginPage.logIn('jsmith@demo.io', 'Demo123!');
-      
-      await LoginPage.clickElement(LoginPage.signInButton);
+      await BasePage.clickElement(LoginPage.signInButton);
+      expect(browser).toHaveUrlContaining('home');
 
     });
 
     it('Should log into demo account when using enter key', async () => {
-      await LoginPage.open('login');
-      LoginPage.logIn('jsmith@demo.io', 'Demo123!');
-
-      let loginButton = $('[id="submit"]');
+      await BasePage.open('login');
+      await LoginPage.logIn('jsmith@demo.io', 'Demo123!');
+      let loginButton = await $('[id="submit"]');
       await loginButton.keys('Enter');
+      expect(browser).toHaveUrlContaining('home');
 
     });
 
     it('Should deny access when invalid data is entered', async () => {
-      await LoginPage.open('login');
-      LoginPage.logIn('luciagonzalez@demo.io', 'Demo123!');
-
-      await LoginPage.clickElement(LoginPage.signInButton);
-
-      let alert = await $('[class= "badge badge-pill badge-danger"]');
-      expect(alert).toHaveText('Error');
+      await BasePage.open('login');
+      await LoginPage.logIn('luciagonzalez@demo.io', 'Demo123!');
+      await BasePage.clickElement(LoginPage.signInButton);
+      expect(LoginPage.alert).toHaveText('Error');
        
     });
 
     it('Should deny access when fields are empty', async () => {
-      await LoginPage.open('login');
-
-      await LoginPage.clickElement(LoginPage.signInButton);
-
+      await BasePage.open('login');
+      await BasePage.clickElement(LoginPage.signInButton);
+      expect(LoginPage.alert).toHaveText('Error');
     });
 
-    it('Should remember the session when the button is checked', async () => {
-      await LoginPage.open('login');
+    it.skip('Should remember the session when the button is checked', async () => {
+      await BasePage.open('login');
       await LoginPage.logIn('jsmith@demo.io', 'Demo123!');
+      await BasePage.clickElement(LoginPage.rememberMe);
+      await BasePage.clickElement(LoginPage.signInButton);
 
-      await LoginPage.clickElement(LoginPage.rememberMe);
+      //tiene que cerrar la pestaña y volver a abrirla
 
-      await LoginPage.clickElement(LoginPage.signInButton);
-
+      //let welcomeMsg = await $('[class="active"]') ;
+      //expect(welcomeMsg).toHaveText('Welcome Josh');
     }); 
 
  });
